Rename commasepstr.py to prog11-15.py
"WAP which takes string as input and display that string in following way:                      Input: Ancit,  Output:  ANCIT ANCI ANC AN A"	 "Write a program which accepts a sequence of comma-separated numbers from console and generate a list and a tuple which contains every number. Suppose the following input is supplied to the program: "	"Input: 34,67,55,33,12,98 Output: ['34', '67', '55', '33', '12', '98'] ('34', '67', '55', '33', '12', '98')" write a program to generate a dictionary that contains (i, i*i) such that is an integral number between 1 and n (both included). and then the program should print the dictionary.	"Input: 8 Output: {1: 1, 2: 4, 3: 9, 4: 16, 5: 25, 6: 36, 7: 49, 8: 64}" "Define a class which has at least two methods: getString: to get a string from console input printString: to print the string in upper case. Also please include simple test function to test the class methods."	 "Write a program that accepts a comma separated sequence of words as input and prints the words in a comma-separated sequence after sorting them alphabetically. Suppose the following input is supplied to the program:"	"Input:  without,hello,bag,world Output: bag,hello,without,world" "Write a program that accepts a sequence of whitespace separated words as input and prints the words after removing all duplicate words and sorting them alphanumerically. Suppose the following input is supplied to the program:"	"Input: hello world and practice makes perfect and hello world again Output: again and hello makes perfect practice world " Write a program that accepts a sentence and calculate the number of letters and digits.	"Input: hello world! 123 Output: LETTERS 10 DIGITS 3 "